# src/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "Docket API",
    "version": "0.1.0"
  },
  "host": "localhost:3000",
  "basePath": "/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/forgot-password": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "",
            "in": "body",
            "name": "ForgotPasswordPostRequestBody",
            "required": true,
            "schema": {
              "properties": {
                "email": {
                  "example": "you@example.com",
                  "format": "email",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "tags": [
          "User"
        ],
        "operationId": "forgotPassword",
        "description": "",
        "summary": "Forgot password"
      }
    },
    "/users/reset-password": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "",
            "in": "body",
            "name": "ResetPasswordPostRequestBody",
            "required": true,
            "schema": {
              "properties": {
                "password": {
                  "example": "s3cr3t",
                  "type": "string"
                },
                "passwordConfirmation": {
                  "example": "s3cr3t",
                  "type": "string"
                },
                "code": {
                  "example": "header.payload.signature",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "tags": [
          "User"
        ],
        "operationId": "resetPassword",
        "description": "",
        "summary": "Reset password"
      }
    },
...
    "/i18n/{localeCode}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "localeCode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "translations": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "tags": [
          "I18n"
        ],
        "operationId": "getLocale",
        "description": "",
        "summary": ""
      }
    },
...
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "example": "An error occurred",
          "type": "string"
        },
        "name": {
          "example": "Error",
          "type": "string"
        },
        "translationKey": {
          "example": "clickHereToLogin",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
...
}